+++ templates/default/lib/app.ex
@@ -4,7 +4,7 @@ defmodule #{projectNamespace} do
 
   def start(_type, _args) do
     {:ok, root} = :file.get_cwd()
-    run_weber(:#{projectName}, Route.get_route, root, Config.config)
+    run_weber(:#{projectName}, Route.__route__, root, Config.config)
   end
 
   def stop(_state) do