+++ templates/default/lib/route.ex
@@ -1,10 +1,8 @@
 defmodule Route do
 
   import Weber.Route
+  require Weber.Route
+  
+  route on("/", :#{projectNamespace}.Main, :action)
 
-  @route on("/", :#{projectNamespace}.Main, :action)
-
-  def get_route do
-    @route
-  end
 end