+++ Changelog.md
@@ -13,6 +13,7 @@
   * `content_for_layout` helper.
   * Controller Helpers.
   * `layout` controller helper.
+  * Get rid of `get_route`, route macros instead.
 
 #### Weber-0.0.2 [Released]